+++ common/src/main/java/traben/entity_model_features/EMFManager.java
@@ -66,22 +66,18 @@ public class EMFManager {//singleton for data holding and resetting needs
     }};
     public final Object2ObjectOpenHashMap<String, EMFJemData> cache_JemDataByFileName = new Object2ObjectOpenHashMap<>();
     public final Object2ObjectOpenHashMap<EMFModel_ID, ModelLayerLocation> cache_LayersByModelName = new Object2ObjectOpenHashMap<>();
+    public final Set<String> EBE_JEMS_FOUND_LAST = new HashSet<>();
     private final Object2IntOpenHashMap<ModelLayerLocation> amountOfLayerAttempts = new Object2IntOpenHashMap<>() {{
         defaultReturnValue(0);
     }};
     private final Set<String> EBE_JEMS_FOUND = new HashSet<>();
-    public final Set<String> EBE_JEMS_FOUND_LAST = new HashSet<>();
+    private final ArrayList<String> KNOWN_RESOURCEPACK_ORDER;
     public UUID entityForDebugPrint = null;
     public long entityRenderCount = 0;
     public boolean isAnimationValidationPhase = false;
     public String currentSpecifiedModelLoading = "";
     public BlockEntityType<?> currentBlockEntityTypeLoading = null;
     private boolean traderLlamaHappened = false;
-
-    public boolean wasEBEModified() {
-        return !EBE_JEMS_FOUND_LAST.isEmpty();
-    }
-
 
 
     private EMFManager() {
@@ -101,19 +97,6 @@ public class EMFManager {//singleton for data holding and resetting needs
         if (self == null) self = new EMFManager();
         return self;
     }
-
-    public ArrayList<String> getResourcePackList() {
-        if (KNOWN_RESOURCEPACK_ORDER.isEmpty()) {
-            var list = Minecraft.getInstance().getResourceManager().listPacks().toList();
-            for (final PackResources pack : list) {
-                this.KNOWN_RESOURCEPACK_ORDER.add(pack.packId());
-            }
-        }
-        return KNOWN_RESOURCEPACK_ORDER;
-    }
-
-    private final ArrayList<String> KNOWN_RESOURCEPACK_ORDER;
-
 
     public static void resetInstance() {
         EMFUtils.log("[EMF (Entity Model Features)]: Clearing data for reload.", false, true);
@@ -185,6 +168,34 @@ public class EMFManager {//singleton for data holding and resetting needs
         return null;
     }
 
+    private static void handleBoats(final String originalLayerName, final EMFModel_ID mobNameForFileAndMap) {
+        String jem;
+        if (originalLayerName.startsWith("chest_boat/")) {
+            jem = originalLayerName.startsWith("chest_boat/bamboo") ? "chest_raft" : "chest_boat";
+        } else if (originalLayerName.startsWith("boat/")) {
+            jem = originalLayerName.startsWith("boat/bamboo") ? "raft" : "boat";
+        } else {
+            return;
+        }
+
+        mobNameForFileAndMap.setMapIdAndAddFallbackModel(jem);
+        String type = mobNameForFileAndMap.getfileName().split("/")[1];
+        mobNameForFileAndMap.setFileName(type + "_" + jem);
+    }
+
+    public boolean wasEBEModified() {
+        return !EBE_JEMS_FOUND_LAST.isEmpty();
+    }
+
+    public ArrayList<String> getResourcePackList() {
+        if (KNOWN_RESOURCEPACK_ORDER.isEmpty()) {
+            var list = Minecraft.getInstance().getResourceManager().listPacks().toList();
+            for (final PackResources pack : list) {
+                this.KNOWN_RESOURCEPACK_ORDER.add(pack.packId());
+            }
+        }
+        return KNOWN_RESOURCEPACK_ORDER;
+    }
 
     public void modifyEBEIfRequired() {
         if (IS_EBE_INSTALLED && !EBE_JEMS_FOUND.isEmpty() && EMF.config().getConfig().allowEBEModConfigModify) {
@@ -215,14 +226,13 @@ public class EMFManager {//singleton for data holding and resetting needs
                         ? originalLayerName : currentSpecifiedModelLoading);
 
 
-
         try {
             EMFManager.lastCreatedRootModelPart = null;
             boolean printing = (EMF.config().getConfig().logModelCreationData);
 
-            if (!"main".equals(layer. #if MC > MC_21 layer() #else getLayer() #endif)) {
-                mobNameForFileAndMap.setBoth(mobNameForFileAndMap.getfileName() + "_" + layer. #if MC > MC_21 layer() #else getLayer() #endif);
-                originalLayerName = originalLayerName + "_" + layer. #if MC > MC_21 layer() #else getLayer() #endif;
+            if (!"main".equals(layer. #if MC > MC_21 layer() #else getLayer() #endif )) {
+                mobNameForFileAndMap.setBoth(mobNameForFileAndMap.getfileName() + "_" + layer. #if MC > MC_21 layer() #else getLayer() #endif );
+                originalLayerName = originalLayerName + "_" + layer. #if MC > MC_21 layer() #else getLayer() #endif ;
             }
 
             boolean modded;
@@ -249,55 +259,53 @@ public class EMFManager {//singleton for data holding and resetting needs
                     case "evoker" -> mobNameForFileAndMap.addFallbackModel("evocation_illager");
                     case "evoker_fangs" -> mobNameForFileAndMap.addFallbackModel("evocation_fangs");
                     case "vindicator" -> mobNameForFileAndMap.addFallbackModel("vindication_illager");
-                    case "bed_foot" -> mobNameForFileAndMap.setBoth("bed", "bed_foot");
-                    case "bed_head" -> mobNameForFileAndMap.setBoth("bed", "bed_head");
+                    case "bed_foot" -> mobNameForFileAndMap.setBoth("bed_foot").addFallbackModel("bed");
+                    case "bed_head" -> mobNameForFileAndMap.setBoth("bed_head").addFallbackModel("bed");
                     case "book" -> {
                         if (currentSpecifiedModelLoading.equals("enchanting_book")) {
-                            mobNameForFileAndMap.setBoth("enchanting_book", "book");
+                            mobNameForFileAndMap.setBoth("enchanting_book", "book").addFallbackModel("book");
                         } else {/* if(currentSpecifiedModelLoading.equals("lectern_book"))*/
-                            mobNameForFileAndMap.setBoth("lectern_book", "book");
+                            mobNameForFileAndMap.setBoth("lectern_book", "book").addFallbackModel("book");
                         }
-                        mobNameForFileAndMap.addFallbackModel("book");
                     }
-                    case "salmon_small", "salmon_large" -> {
-//                        mobNameForFileAndMap.setRootTransformer(null);
-                        mobNameForFileAndMap.addFallbackModel( "salmon");//getModelScalingTransformer(layer));
-                    }
+                    case "salmon_small", "salmon_large" -> mobNameForFileAndMap.addFallbackModel("salmon");
                     case "breeze_wind_charge" -> mobNameForFileAndMap.setMapIdAndAddFallbackModel("wind_charge");
                     case "creaking_transient" -> mobNameForFileAndMap.setMapIdAndAddFallbackModel("creaking");
                     case "chest" -> mobNameForFileAndMap.setBoth(currentSpecifiedModelLoading, "chest");
-                    case "conduit_cage" -> mobNameForFileAndMap.setBoth("conduit", "conduit_cage");
-                    case "conduit_eye" -> mobNameForFileAndMap.setBoth("conduit", "conduit_eye");
-                    case "conduit_shell" -> mobNameForFileAndMap.setBoth("conduit", "conduit_shell");
-                    case "conduit_wind" -> mobNameForFileAndMap.setBoth("conduit", "conduit_wind");
+                    case "conduit_cage" -> mobNameForFileAndMap.setBoth("conduit_cage").addFallbackModel("conduit");
+                    case "conduit_eye" -> mobNameForFileAndMap.setBoth("conduit_eye").addFallbackModel("conduit");
+                    case "conduit_shell" -> mobNameForFileAndMap.setBoth("conduit_shell").addFallbackModel("conduit");
+                    case "conduit_wind" -> mobNameForFileAndMap.setBoth("conduit_wind").addFallbackModel("conduit");
                     case "creeper_armor" -> mobNameForFileAndMap.setBoth("creeper_charge");
                     case "creeper_head" -> mobNameForFileAndMap.setBoth("head_creeper");
-                    case "decorated_pot_base" -> mobNameForFileAndMap.setBoth("decorated_pot", "decorated_pot_base");
-                    case "decorated_pot_sides" -> mobNameForFileAndMap.setBoth("decorated_pot", "decorated_pot_sides");
+                    case "decorated_pot_base" ->
+                            mobNameForFileAndMap.setBoth("decorated_pot_base").addFallbackModel("decorated_pot");
+                    case "decorated_pot_sides" ->
+                            mobNameForFileAndMap.setBoth("decorated_pot_sides").addFallbackModel("decorated_pot");
                     case "double_chest_left" -> getDoubleChest(root, mobNameForFileAndMap, false, printing);
                     case "double_chest_right" -> getDoubleChest(root, mobNameForFileAndMap, true, printing);
                     case "dragon_skull" -> mobNameForFileAndMap.setBoth("head_dragon");
                     case "ender_dragon" -> mobNameForFileAndMap.setBoth("dragon");
                     case "leash_knot" -> mobNameForFileAndMap.setBoth("lead_knot");
                     case "llama", "llama_baby" -> traderLlamaHappened = false;
-                    case "llama_decor" -> mobNameForFileAndMap.setBoth(traderLlamaHappened ? "trader_llama_decor" : "llama_decor");
-                    case "llama_baby_decor" -> mobNameForFileAndMap.setBoth(traderLlamaHappened ? "trader_llama_baby_decor" : "llama_baby_decor");
-                    case "chest_minecart", "command_block_minecart", "spawner_minecart", "tnt_minecart", "furnace_minecart", "hopper_minecart"
-                            -> mobNameForFileAndMap.setBoth("minecart");
+                    case "llama_decor" ->
+                            mobNameForFileAndMap.setBoth(traderLlamaHappened ? "trader_llama_decor" : "llama_decor");
+                    case "llama_baby_decor" ->
+                            mobNameForFileAndMap.setBoth(traderLlamaHappened ? "trader_llama_baby_decor" : "llama_baby_decor");
+                    case "chest_minecart", "command_block_minecart", "spawner_minecart", "tnt_minecart",
+                         "furnace_minecart", "hopper_minecart" -> mobNameForFileAndMap.setBoth("minecart");
                     case "piglin_head" -> mobNameForFileAndMap.setBoth("head_piglin");
                     case "player_head" -> mobNameForFileAndMap.setBoth("head_player");
+                    case "player_slim" -> mobNameForFileAndMap.addFallbackModel("player");
                     case "boat_water_patch" -> {
-                        if(currentSpecifiedModelLoading.startsWith("emf$boat$")){
+                        if (currentSpecifiedModelLoading.startsWith("emf$boat$")) {
                             String type = currentSpecifiedModelLoading.substring(9);
-                            mobNameForFileAndMap.setBoth(type + "_boat_patch","boat_patch");
+                            mobNameForFileAndMap.setBoth(type + "_boat_patch", "boat_patch").addFallbackModel("boat_patch");
                             currentSpecifiedModelLoading = "";
-                        }else{
+                        } else {
                             mobNameForFileAndMap.setBoth("boat_patch");
                         }
-
-                        mobNameForFileAndMap.addFallbackModel("boat_patch");
                     }
-
                     case "pufferfish_big" -> mobNameForFileAndMap.setBoth("puffer_fish_big");
                     case "pufferfish_medium" -> mobNameForFileAndMap.setBoth("puffer_fish_medium");
                     case "pufferfish_small" -> mobNameForFileAndMap.setBoth("puffer_fish_small");
@@ -330,25 +338,25 @@ public class EMFManager {//singleton for data holding and resetting needs
 
                                     if (originalLayerName.startsWith("sign/standing/")) {
                                         sign += "_sign";
-                                        mobNameForFileAndMap.setFileName(sign);
-                                        mobNameForFileAndMap.setMapIdAndAddFallbackModel("sign");
+                                        mobNameForFileAndMap.setFileName(sign)
+                                                .setMapIdAndAddFallbackModel("sign");
                                     } else if (originalLayerName.startsWith("sign/wall/")) {
                                         sign += "_wall_sign";
-                                        mobNameForFileAndMap.setFileName(sign);
-                                        mobNameForFileAndMap.setMapIdAndAddFallbackModel("wall_sign");
-                                        mobNameForFileAndMap.setMapIdAndAddFallbackModel("sign");
+                                        mobNameForFileAndMap.setFileName(sign)
+                                                .setMapIdAndAddFallbackModel("wall_sign")
+                                                .setMapIdAndAddFallbackModel("sign");
                                     } else {
                                         sign += "_hanging_sign";
-                                        mobNameForFileAndMap.setFileName(sign);
-                                        mobNameForFileAndMap.setMapIdAndAddFallbackModel("hanging_sign");
+                                        mobNameForFileAndMap.setFileName(sign)
+                                                .setMapIdAndAddFallbackModel("hanging_sign");
                                     }
                                 }
                                 default -> {
 
                                     if (EMF.config().getConfig().modelExportMode != EMFConfig.ModelPrintMode.NONE)
                                         EMFUtils.log("EMF unknown modifiable block entity model identified during loading: " + currentSpecifiedModelLoading + ".jem");
-                                    mobNameForFileAndMap.setFileName(currentSpecifiedModelLoading);
-                                    mobNameForFileAndMap.setMapIdAndAddFallbackModel(currentSpecifiedModelLoading,originalLayerName);
+                                    mobNameForFileAndMap.setFileName(currentSpecifiedModelLoading)
+                                            .setMapIdAndAddFallbackModel(currentSpecifiedModelLoading, originalLayerName);
                                 }
                             }
                         } else if (originalLayerName.contains("/") && layer. #if MC > MC_21 layer() #else getLayer() #endif .equals("main")) {
@@ -375,19 +383,15 @@ public class EMFManager {//singleton for data holding and resetting needs
 
             //cache the layers for the model
             cache_LayersByModelName.put(mobNameForFileAndMap, layer);
-            var fallbackModel = mobNameForFileAndMap;
-            while (fallbackModel.hasFallbackModels()) {
-                fallbackModel = fallbackModel.getNextFallbackModel();
-                if (fallbackModel == null) break;
-                cache_LayersByModelName.putIfAbsent(fallbackModel, layer);
-            }
+            mobNameForFileAndMap.forEachFallback((fallBack) -> cache_LayersByModelName.putIfAbsent(fallBack, layer));
 
-            ///jem name and secondaries are final and correct from here
+            ///jem name and fallbacks are final and correct from here
 
 
-            if (printing) EMFUtils.log(" > checking if: [" + mobNameForFileAndMap +"], is allowed as a model name.");
-            if(EMF.config().getConfig().isModelDisabled(mobNameForFileAndMap.getMapId())) {
-                if (printing) EMFUtils.logWarn(" > Vanilla model used for: [" + mobNameForFileAndMap + "], because it is disabled in EMF's settings or via the API.");//original name
+            if (printing) EMFUtils.log(" > checking if: [" + mobNameForFileAndMap + "], is allowed as a model name.");
+            if (EMF.config().getConfig().isModelDisabled(mobNameForFileAndMap.getMapId())) {
+                if (printing)
+                    EMFUtils.logWarn(" > Vanilla model used for: [" + mobNameForFileAndMap + "], because it is disabled in EMF's settings or via the API.");//original name
                 ((IEMFModelNameContainer) root).emf$insertKnownMappings(mobNameForFileAndMap);
                 return root;
             }
@@ -395,18 +399,17 @@ public class EMFManager {//singleton for data holding and resetting needs
             //construct simple map for modded or unknown entities
             Map<String, String> optifinePartNameMap = EMFModelMappings.getMapOf(mobNameForFileAndMap.getMapId(), root);
 
-            if (printing) EMFUtils.log(" >> EMF trying to find model: " + mobNameForFileAndMap.getNamespace() + ":optifine/cem/" + mobNameForFileAndMap + ".jem");
+            if (printing)
+                EMFUtils.log(" >> EMF trying to find model: " + mobNameForFileAndMap.getNamespace() + ":optifine/cem/" + mobNameForFileAndMap + ".jem");
 
             var modelDataAndContext = getJemAndContext(printing, mobNameForFileAndMap);
 
             //try with secondary model
 
-            var fallbackModelId = mobNameForFileAndMap;
-            while(fallbackModelId.hasFallbackModels()){
-                fallbackModelId = fallbackModelId.getNextFallbackModel();
-                if(fallbackModelId == null) throw new Exception("Fallback model is null");
-
-                if (printing) EMFUtils.log(" >> EMF trying to find fallback model: " + fallbackModelId.getNamespace() + ":optifine/cem/" + fallbackModelId + ".jem");
+//
+            mobNameForFileAndMap.forEachFallback((fallbackModelId) -> {
+                if (printing)
+                    EMFUtils.log(" >> EMF trying to find fallback model: " + fallbackModelId.getNamespace() + ":optifine/cem/" + fallbackModelId + ".jem");
 
                 var fallbackModelDataAndContext = getJemAndContext(printing, fallbackModelId);
 
@@ -421,7 +424,7 @@ public class EMFManager {//singleton for data holding and resetting needs
                     modelDataAndContext.setMiddle(fallbackModelDataAndContext.getMiddle());
                     modelDataAndContext.setRight(fallbackModelId);
                 }
-            }
+            });
 
             EMFJemData jemData = modelDataAndContext.getLeft();
             var directoryContextBaseAndVariant = modelDataAndContext.getMiddle();
@@ -433,9 +436,9 @@ public class EMFManager {//singleton for data holding and resetting needs
                 //we do have custom models
 
                 //abort with message if we have variant models and no base model and the setting to require this like OptiFine is set
-                if(jemData == null && EMF.config().getConfig().enforceOptifineVariationRequiresDefaultModel){
-                    EMFUtils.logWarn("The model [" + finalMapData.getfileName() +"] has variation but does not have a default 'base' model, this is not allowed in the OptiFine format.\nYou may disable this requirement in EMF in the 'model > options' settings. Though it is usually best to preserve OptiFine compatibility.\nYou can get a default model by exporting it in the EMF settings via 'models > allmodels > *model* > export'");
-                }else {
+                if (jemData == null && EMF.config().getConfig().enforceOptifineVariationRequiresDefaultModel) {
+                    EMFUtils.logWarn("The model [" + finalMapData.getfileName() + "] has variation but does not have a default 'base' model, this is not allowed in the OptiFine format.\nYou may disable this requirement in EMF in the 'model > options' settings. Though it is usually best to preserve OptiFine compatibility.\nYou can get a default model by exporting it in the EMF settings via 'models > allmodels > *model* > export'");
+                } else {
                     //specification for the optifine map
                     // only used for tadpole head parts currently as optifine actually uses the root as the body
                     Set<String> optifinePartNames = new HashSet<>();
@@ -445,7 +448,7 @@ public class EMFManager {//singleton for data holding and resetting needs
                         }
                     });
 
-                    EMFModelPartRoot emfRoot = new EMFModelPartRoot(finalMapData, Objects.requireNonNullElseGet(directoryContextBaseAndVariant.getLeft(),directoryContextBaseAndVariant::getRight), root, optifinePartNames, new HashMap<>());
+                    EMFModelPartRoot emfRoot = new EMFModelPartRoot(finalMapData, Objects.requireNonNullElseGet(directoryContextBaseAndVariant.getLeft(), directoryContextBaseAndVariant::getRight), root, optifinePartNames, new HashMap<>());
                     if (jemData != null) {
                         emfRoot.addVariantOfJem(jemData, 1);
                         emfRoot.setVariantStateTo(1);
@@ -453,8 +456,8 @@ public class EMFManager {//singleton for data holding and resetting needs
                         emfRoot.containsCustomModel = true;
                         if (hasVariants) {
                             emfRoot.discoverAndInitVariants();
-                        }else if (!modded && jemData.directoryContext.isSubFolder && EMF.config().getConfig().enforceOptifineSubFoldersVariantOnly){
-                            EMFUtils.logError("Error loading ["+jemData.directoryContext.getFinalFileLocation()+"] as it is in a subfolder but does not have any variants. This is not allowed in the OptiFine format. You may disable this requirement in EMF's settings at 'model > OptiFine settings'. Though it is usually best to preserve OptiFine compatibility.");
+                        } else if (!modded && jemData.directoryContext.isSubFolder && EMF.config().getConfig().enforceOptifineSubFoldersVariantOnly) {
+                            EMFUtils.logError("Error loading [" + jemData.directoryContext.getFinalFileLocation() + "] as it is in a subfolder but does not have any variants. This is not allowed in the OptiFine format. You may disable this requirement in EMF's settings at 'model > OptiFine settings'. Though it is usually best to preserve OptiFine compatibility.");
                             throw new Exception("Subfolder without variants, OptiFine compat enabled");
                         }
                     } else {
@@ -491,34 +494,19 @@ public class EMFManager {//singleton for data holding and resetting needs
         }
     }
 
-    private static void handleBoats(final String originalLayerName, final EMFModel_ID mobNameForFileAndMap) {
-        String jem;
-        if (originalLayerName.startsWith("chest_boat/")) {
-            jem = originalLayerName.startsWith("chest_boat/bamboo") ? "chest_raft" : "chest_boat";
-        } else if (originalLayerName.startsWith("boat/")) {
-            jem = originalLayerName.startsWith("boat/bamboo") ? "raft" : "boat";
-        } else {
-            return;
-        }
-
-        mobNameForFileAndMap.setMapIdAndAddFallbackModel(jem);
-        String type = mobNameForFileAndMap.getfileName().split("/")[1];
-        mobNameForFileAndMap.setFileName(type + "_" + jem);
-    }
-
-    private MutableTriple<EMFJemData, ImmutablePair<EMFDirectoryHandler, EMFDirectoryHandler>, EMFModel_ID> getJemAndContext(boolean printing, EMFModel_ID mobNameForFileAndMap){
+    private MutableTriple<EMFJemData, ImmutablePair<EMFDirectoryHandler, EMFDirectoryHandler>, EMFModel_ID> getJemAndContext(boolean printing, EMFModel_ID mobNameForFileAndMap) {
         EMFDirectoryHandler baseModelDir = EMFDirectoryHandler.getDirectoryManagerOrNull(printing, mobNameForFileAndMap.getNamespace(), mobNameForFileAndMap.getfileName(), ".jem");
         EMFDirectoryHandler propertiesOrSecondDir = EMFDirectoryHandler.getDirectoryManagerOrNull(printing, mobNameForFileAndMap.getNamespace(), mobNameForFileAndMap.getfileName(), ".properties");
         if (propertiesOrSecondDir == null)
             propertiesOrSecondDir = EMFDirectoryHandler.getDirectoryManagerOrNull(printing, mobNameForFileAndMap.getNamespace(), mobNameForFileAndMap.getfileName(), "2.jem");
 
         //discard the variation context if they are in a lower pack to the base model and not in matching sub folder locations
-        if(baseModelDir != null && !baseModelDir.validForThisBase(propertiesOrSecondDir)) propertiesOrSecondDir = null;
+        if (baseModelDir != null && !baseModelDir.validForThisBase(propertiesOrSecondDir)) propertiesOrSecondDir = null;
 
         EMFJemData jemDataFirst = baseModelDir == null ? null : getJemDataWithDirectory(baseModelDir, mobNameForFileAndMap);
 
         //pack result
-        return MutableTriple.of(jemDataFirst, ImmutablePair.of(baseModelDir,propertiesOrSecondDir),mobNameForFileAndMap);
+        return MutableTriple.of(jemDataFirst, ImmutablePair.of(baseModelDir, propertiesOrSecondDir), mobNameForFileAndMap);
     }
 
     private void getDoubleChest(ModelPart root, EMFModel_ID mobNameForFileAndMap, boolean isRight, boolean printing) {
@@ -527,14 +515,15 @@ public class EMFManager {//singleton for data holding and resetting needs
 
         mobNameForFileAndMap.setBoth(currentSpecifiedModelLoading + "_large", "double_chest_" + thisSide);
         if (EMF.config().getConfig().doubleChestAnimFix) {
-            if (printing) EMFUtils.log("injecting empty " + otherSide + " side parts into 'double chest' for animation purposes");
+            if (printing)
+                EMFUtils.log("injecting empty " + otherSide + " side parts into 'double chest' for animation purposes");
             Map<String, ModelPart> newChildren = new HashMap<>(root.children);
             newChildren.putIfAbsent("lid_" + otherSide, new ModelPart(List.of(), Map.of()));
             newChildren.putIfAbsent("base_" + otherSide, new ModelPart(List.of(), Map.of()));
             newChildren.putIfAbsent("knob_" + otherSide, new ModelPart(List.of(), Map.of()));
             root.children = newChildren; // mutable
         }
-        mobNameForFileAndMap.addFallbackModel(mobNameForFileAndMap.namespace,mobNameForFileAndMap.getDisplayFileName());
+        mobNameForFileAndMap.addFallbackModel(mobNameForFileAndMap.namespace, mobNameForFileAndMap.getDisplayFileName());
         mobNameForFileAndMap.setFileName(currentSpecifiedModelLoading + "_" + thisSide);
     }